pyproject.toml
[tool.poetry]
name = "desk-gpt"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]
license = "Unlicense"
readme = "README.md"
packages = [{include = "desk_gpt"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
python-decouple = "^3.8"
openai = "^1.30.1"
django-cors-headers = "^4.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
